+++ lib/get_navigation/src/routes/route_middleware.dart
@@ -198,7 +198,7 @@ class PageRedirect {
     final _r = (isUnknown ? unknownRoute : route)!;
     return GetPageRoute<T>(
       page: _r.page,
-      parameter: _r.parameter,
+      parameter: _r.parameters,
       settings: isUnknown
           ? RouteSettings(
               name: _r.name,
@@ -227,7 +227,7 @@ class PageRedirect {
 
     return GetPageRoute<T>(
       page: _r.page,
-      parameter: _r.parameter,
+      parameter: _r.parameters,
       alignment: _r.alignment,
       title: _r.title,
       maintainState: _r.maintainState,
@@ -279,10 +279,10 @@ class PageRedirect {
   }
 
   void addPageParameter(GetPage route) {
-    if (route.parameter == null) return;
+    if (route.parameters == null) return;
 
     final parameters = Get.parameters;
-    parameters.addEntries(route.parameter!.entries);
+    parameters.addEntries(route.parameters!.entries);
     Get.parameters = parameters;
   }
 }