+++ lib/get_state_manager/src/rx_flutter/rx_ticket_provider_mixin.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/scheduler.dart';
 import '../../get_state_manager.dart';
 
-/// Used like [SingleTickerProviderMixin] but only with Get Controllers.
+/// Used like `SingleTickerProviderMixin` but only with Get Controllers.
 /// Simplifies AnimationController creation inside GetxController.
 ///
 /// Example: